Standardize test file organization and naming across all apps.
