+++ keystone_api/apps/health/tests/test_views/test_HealthCheckPrometheusView.py
@@ -2,7 +2,7 @@
 
 from django.test import TestCase
 
-from apps.health.tests.views.utils import create_mock_plugin
+from apps.health.tests.test_views.utils import create_mock_plugin
 from apps.health.views import HealthCheckPrometheusView
