+++ keystone_api/apps/notifications/tests/test_shortcuts/test_send_notification_template.py
